pyproject.toml
[tool.poetry]
name = "fastapi-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fastapi_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
httpx = "^0.27.0"
jwt = "^1.3.1"
msal = "^1.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
azure-cosmos = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
